% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_orders_chrono}
\alias{all_orders_chrono}
\title{Open Tree of Life Orders}
\format{data frame}
\source{
\url{}
}
\usage{
all_orders_chrono
}
\description{
Data Frame of all order in Open Tree of Life Taxonomy
}
\details{
Generated with
data(all_ott_orders)
all_orders_chrono <- datelife::get_dated_otol_induced_subtree(ott_id = all_ott_orders$ott_id)
missing_index <- is.na(match(all_ott_orders$name, all_orders_chrono$tip.label))
missing_ott_orders <- all_ott_orders$ott_id[missing_index]
names(missing_ott_orders) <- all_ott_orders$name[missing_index]
all_orders_chrono$missing_ott_orders <- missing_ott_orders
usethis::use_data(all_orders_chrono, overwrite = TRUE)
}
\keyword{chronogram}
\keyword{orders}
\keyword{otol}
