% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_ott_orders}
\alias{all_ott_orders}
\title{Open Tree of Life Orders}
\format{data frame}
\source{
\url{}
}
\usage{
all_ott_orders
}
\description{
Data Frame of all order in Open Tree of Life Taxonomy
}
\details{
Generated with
grep -i "order" taxonomy.tsv > orders.tsv
grep -i "order" taxonomy.tsv | grep -v "genus" | grep -v "no rank - terminal" | grep -v "merged" | grep -v "barren" | grep -v "extinct" | grep -v "major_rank_conflict_inherited" | grep -v "no rank" | grep -v "family" | grep -v "suborder" | grep -v "species" | grep -v "infraorder" | grep -v "unplaced" | grep -v "superorder" | grep -v "tribe" | grep -v "parvorder" > orders.tsv
all_ott_orders <- read.table("/Users/luna/GoogleDrive/datelife/otol/taxonomy/ott/orders.tsv", sep = "|", header = FALSE, fill = TRUE, stringsAsFactors=FALSE)
names(all_ott_orders) <- c("ott_id", "parent_ott_id", "name", "rank", "source_info", "unique_name", "flags", "?")
for (i in c("name", "rank", "source_info", "unique_name", "flags")){
  all_ott_orders[,i] <- gsub("\\t", "", all_ott_orders[,i])
}
all_ott_orders[,"source_info"] <- gsub("/", "", all_ott_orders[,"source_info"])
unique(all_ott_orders[,"flags"])
usethis::use_data(all_ott_orders, overwrite = TRUE)
}
\keyword{orders}
\keyword{otol}
\keyword{ott}
