% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funded_taxa_clean.R
\name{clean_tnrs}
\alias{clean_tnrs}
\title{it eliminates unmatched (NAs) and invalid taxa from a tnrs_match_names object
returns a data frame}
\usage{
clean_tnrs(tnrs)
}
\description{
it eliminates unmatched (NAs) and invalid taxa from a tnrs_match_names object
returns a data frame
}
