% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/get_family_tree.R
\docType{data}
\name{fam_tree}
\alias{fam_tree}
\alias{plant_tree_orders}
\title{Open Tree of Life Land Plant Orders Chronogram}
\format{A phylo object.}
\source{
\url{http://www.otol.org}

\url{http://www.otol.org}
}
\usage{
plant_tree_orders

fam_tree
}
\description{
Dated induced OToL subtree of OTT order rank taxa of Land Plants

Induced OTL subtree of family rank taxa from OTT
}
\details{
Generated with code in data-raw/plant_family_tree-2dateA.R

Generated with fam_tree <- get_otol_family_tree()
}
\keyword{chronogram}
\keyword{families}
\keyword{land}
\keyword{orders}
\keyword{otol}
\keyword{plants}
\keyword{time}
\keyword{tree}
