% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funded_taxa.R
\name{get_funded_taxa}
\alias{get_funded_taxa}
\title{Get taxa from awarded grant proposals within Systematics and Phylogenetics}
\usage{
get_funded_taxa(database = "grants.rda")
}
\arguments{
\item{database}{A character vector with the database to get taxa from. Default to rnsf package grants database}
}
\value{
A data frame with grant data
}
\description{
Get taxa from awarded grant proposals within Systematics and Phylogenetics
}
