% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family_tree.R
\name{get_otol_family_tree}
\alias{get_otol_family_tree}
\title{function to get a family tree level from Open Tree of Life synthetic tree}
\usage{
get_otol_family_tree()
}
\description{
function to get a family tree level from Open Tree of Life synthetic tree
}
