% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funded_taxa_clean.R
\name{get_valid_children_names}
\alias{get_valid_children_names}
\title{identifies valid children names, ott_ids and ranks from a taxonomy_taxon_info output
returns a list with valid children unique OTT names, ott_ids and ranks}
\usage{
get_valid_children_names(taxon_info)
}
\description{
identifies valid children names, ott_ids and ranks from a taxonomy_taxon_info output
returns a list with valid children unique OTT names, ott_ids and ranks
}
