% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funded_taxa.R
\docType{data}
\name{grants}
\alias{grants}
\title{NSF awards information database from rnsf package}
\format{A list of 45 elements, with information on NSF awarded grants.}
\source{
\url{http://www.nsf.org}
}
\usage{
grants
}
\description{
It includes grants from its start until the databse was last copied in 2018.10.15.
}
\keyword{grant}
\keyword{nsf}
