% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funded_taxa.R
\docType{data}
\name{nsf_relevant_grants_raw}
\alias{nsf_relevant_grants_raw}
\title{Systematics and Phylogenetics NSF awards information and studied taxa}
\format{A list of 49 elements, with infromstion on NSF awarded grants in Systematics and Phylogenetics.
\describe{
  \item{abstractText}{A character vector of grant abstracts}
  \item{agency}{A character vector of }
  \item{awardAgencyCode}{A character vector of }
  \item{awardeeAddress}{A character vector of }
  \item{awardeeCity}{A character vector of }
  \item{awardeeCountryCode}{A character vector of }
  \item{awardeeCounty}{A character vector of }
  \item{awardeeDistrictCode}{A character vector of }
  \item{awardeeName}{A character vector of }
  \item{awardeeStateCode}{A character vector of }
  \item{awardeeZipCode}{A character vector of }
  \item{cfdaNumber}{A character vector of }
  \item{dunsNumber}{A character vector of }
  \item{estimatedTotalAmt}{A character vector of }
  \item{fundsObligatedAmt}{A character vector of }
  \item{fundAgencyCode}{A character vector of }
  \item{fundProgramName}{A character vector of }
  \item{id}{A character vector of }
  \item{parentDunsNumber}{A character vector of }
  \item{pdPIName}{A character vector of }
  \item{perfAddress}{A character vector of }
  \item{perfCity}{A character vector of }
  \item{perfCountryCode}{A character vector of }
  \item{perfCounty}{A character vector of }
  \item{perfDistrictCode}{A character vector of }
  \item{perfLocation}{A character vector of }
  \item{perfStateCode}{A character vector of }
  \item{perfZipCode}{A character vector of }
  \item{piFirstName}{A character vector of }
  \item{piLastName}{A character vector of }
  \item{poName}{A character vector of }
  \item{primaryProgram}{A character vector of }
  \item{date}{A character vector of }
  \item{startDate}{A character vector of }
  \item{expDate}{A character vector of }
  \item{title}{A character vector of }
  \item{transType}{A character vector of }
  \item{awardee}{A character vector of }
  \item{coPDPI}{A list of }
  \item{piMiddeInitial}{A character vector of Principal Investigator middle initials}
  \item{publicationResearch}{A list of }
  \item{projectOutComesReport}{A character vector of }
  \item{response.serviceNotification.notificationType}{A character vector of }
  \item{response.serviceNotification.notificationCode}{A character vector of }
  \item{response.serviceNotification.notificationMessage}{A character vector of }
  \item{taxa}{A list of }
  \item{taxa_correct}{A list of }
  \item{name_tnrs}{A list of }
  \item{fam_tnrs}{A list of }
}}
\source{
\url{http://www.nsf.org}
}
\usage{
nsf_relevant_grants_raw
}
\description{
It includes grants from its (NSF) start until the package was last updated.
}
\details{
Generated with nsf_relevant_grants_raw <- get_funded_taxa()
}
\keyword{grant}
\keyword{nsf}
\keyword{phylogenetics}
\keyword{systematics}
\keyword{taxa}
