% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family_tree.R
\docType{data}
\name{fam_tree}
\alias{fam_tree}
\title{Open Tree of Life Family Tree}
\format{A phylo object.}
\source{
\url{http://www.otol.org}
}
\usage{
fam_tree
}
\description{
Induced OTL subtree of family rank taxa from OTT
}
\details{
Generated with fam_tree <- get_otol_family_tree()
}
\keyword{families}
\keyword{otol}
\keyword{tree}
